+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 
+## v1.63.0-rc.18 (2025-09-18)
+
+#### :bug: Bug Fix
+* `recurring-payments-job`
+  * [#2207](https://github.com/DEFRA/rod-licensing/pull/2207) Refactor for Airbrake integration ([@jaucourt](https://github.com/jaucourt))
+
+#### Committers: 1
+- Phil Benson ([@jaucourt](https://github.com/jaucourt))
+
+## v1.63.0-rc.17 (2025-09-16)
+
+#### :bug: Bug Fix
+* `gafl-webapp-service`
+  * [#2210](https://github.com/DEFRA/rod-licensing/pull/2210) Country capital letters ([@ScottDormand96](https://github.com/ScottDormand96))
+
+#### Committers: 1
+- Scott Dormand ([@ScottDormand96](https://github.com/ScottDormand96))
+
 ## v1.63.0-rc.16 (2025-09-09)
 
 #### :bug: Bug Fix